Moved extInstallationLinks to extension folder
